Potential sync fix
Following the change to the POS rate on January 8th it has become
apparent that this has caused problems with regards to syncing the
wallet as it now tries to evaluate already accepted blocks under the
revised rate.

This update uses the specific block number that the change occurred
(242825) instead of the January 8th timestamp and should hopefully
resolve the issue going forward.